+++ implicit_solver/objects/components/constraints/bending.py
@@ -24,7 +24,7 @@ class Bending(Base):
         node_ids[0] - node_ids[1] - node_ids[2]
         '''
         Base.__init__(self, num_nodes = 3)
-        self.rest_angle = np.float32(0.0)
+        self.rest_angle = np.float64(0.0)
 
     def set_object(self, scene, node_ids):
         '''
