+++ implicit_solver/objects/components/node.py
@@ -10,9 +10,9 @@ class Node:
     Describes the constraint base
     '''
     def __init__(self):
-        self.x = np.zeros(2, dtype = np.float32)
-        self.v = np.zeros(2, dtype = np.float32)
-        self.f = np.zeros(2, dtype = np.float32)
-        self.m = np.float32(0.0)
-        self.im = np.float32(0.0)
+        self.x = np.zeros(2, dtype = np.float64)
+        self.v = np.zeros(2, dtype = np.float64)
+        self.f = np.zeros(2, dtype = np.float64)
+        self.m = np.float64(0.0)
+        self.im = np.float64(0.0)
